change Armor scope from player to Character (#34576)
* change scope of armor functions

* move character functions to proper place

* remove auto
